Modified maxpool kernel

+++ kernels/maxpool/include/defs_maxpool.h
@@ -15,11 +15,11 @@ static inline uint32_t read_mcycle(void) {
     (ceil_mode ? ((in_dim) + (stride) - (kernel) + (stride) - 1) / (stride) : ((in_dim) - (kernel)) / (stride) + 1)
 
 // Function prototypes for MaxPool kernels
-void maxpool_scalar(const int32_t* X, int32_t* Y, int32_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
-void maxpool_e32m1(const int32_t* X, int32_t* Y, int32_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
-void maxpool_e32m2(const int32_t* X, int32_t* Y, int32_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
-void maxpool_e32m4(const int32_t* X, int32_t* Y, int32_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
-void maxpool_e32m8(const int32_t* X, int32_t* Y, int32_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
+void maxpool_scalar(const int32_t* X, int32_t* Y, int64_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
+void maxpool_e32m1(const int32_t* X, int32_t* Y, int64_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
+void maxpool_e32m2(const int32_t* X, int32_t* Y, int64_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
+void maxpool_e32m4(const int32_t* X, int32_t* Y, int64_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
+void maxpool_e32m8(const int32_t* X, int32_t* Y, int64_t* I, size_t N, size_t C, size_t H, size_t W, size_t K, size_t S, bool ceil_mode);
 
 // Utility function prototypes
 void read_tensor_binary(const char* filename, float* tensor, size_t count);
